+++ src/test/java/com/example/AlexGetFriendsAndPlaceOfLivingTest.java
@@ -10,7 +10,7 @@ import java.util.List;
 import static org.junit.Assert.assertEquals;
 
 @RunWith(MockitoJUnitRunner.class)
-public class AlexGetFriendsAndPlaceOfLivingTests {
+public class AlexGetFriendsAndPlaceOfLivingTest {
 
     @Mock
     Feline feline;
