+++ src/test/java/com/example/FelineTest.java
@@ -10,7 +10,7 @@ import java.util.List;
 import static org.junit.Assert.assertEquals;
 
 @RunWith(MockitoJUnitRunner.class)
-public class FelineTests {
+public class FelineTest {
 
     @Test
     public void getFamilySuccess(){
