+++ src/test/java/com/example/LionCreateExceptionTest.java
@@ -4,7 +4,7 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
-public class LionCreateExceptionTests {
+public class LionCreateExceptionTest {
 
     Feline feline = new Feline();
